requirements2.txt
aiohttp==3.7.4.post0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
bcrypt==3.2.0
bleach==4.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
discord.py==1.7.2
entrypoints==0.4
Flask==2.0.0
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==4.8.3
imutils==0.5.4
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.0.0
jedi==0.17.2
Jinja2==3.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
kiwisolver==1.3.1
MarkupSafe==2.0.0
matplotlib==3.3.4
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.19.5
opencv-python==4.6.0.66
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.12.0
pyjokes==0.6.0
pyparsing==3.0.9
pyrsistent==0.18.0
pytesseract==0.3.8
python-dateutil==2.8.2
python-dotenv==0.17.1
pytz==2022.1
pywin32==304
pywinpty==1.1.6
pyzmq==23.2.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.25.1
Send2Trash==1.8.0
six==1.16.0
SQLAlchemy==1.4.15
terminado==0.12.1
testpath==0.6.0
toml==0.10.2
tornado==6.1
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.0
widgetsnbextension==3.6.1
WTForms==2.3.3
yarl==1.6.3
zipp==3.6.0
